+++ cpp/src/reductions/sum_of_squares.cu
@@ -16,7 +16,8 @@
 // The translation unit for reduction `sum of squares`
 
 #include <cudf/detail/reduction_functions.hpp>
-#include "simple.cuh"
+#include <cudf/dictionary/dictionary_column_view.hpp>
+#include <reductions/simple.cuh>
 
 std::unique_ptr<cudf::scalar> cudf::reduction::sum_of_squares(column_view const& col,
                                                               cudf::data_type const output_dtype,
@@ -25,5 +26,7 @@ std::unique_ptr<cudf::scalar> cudf::reduction::sum_of_squares(column_view const&
 {
   using reducer =
     cudf::reduction::simple::element_type_dispatcher<cudf::reduction::op::sum_of_squares>;
-  return cudf::type_dispatcher(col.type(), reducer(), col, output_dtype, mr, stream);
+  auto col_type =
+    cudf::is_dictionary(col.type()) ? dictionary_column_view(col).keys().type() : col.type();
+  return cudf::type_dispatcher(col_type, reducer(), col, output_dtype, mr, stream);
 }